updation in frontend and Backend

+++ Backend/src/main/java/com/ankit/crud/Repository/TechStackRepository.java
@@ -3,6 +3,6 @@ package com.ankit.crud.Repository;
 import com.ankit.crud.model.TechStack;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TechStackRepository extends JpaRepository<TechStack, Long> {
+public interface TechStackRepository extends JpaRepository<TechStack, Integer> {
 
 }
